Implement trapezoid-to-rectangle transformation
It's now OK to load a photo of a photo that happened to be taken at an angle,
turn off the "constrained" setting on the menu, drag the box corners to fit
the image corners, and save the cropped image. It will be transformed back
to a head-on rectangular image as if the picture had been taken that way.