Keep track of which rectangles are constrained; read and write xml property
The checkable menu item controls the tool behavior, but the Rectangle
remembers whether or not that was checked at the time it was last adjusted.
So MainWindow::on_actionSave_triggered() obeys that property rather than
the menu checked state. So it's ok to load an xml file and immediately
save out the cropped images.